refactor(core): move LLM drivers to provider crates with DriverRegistry pattern
- Create DriverRegistry in everruns-core for dynamic driver registration
- Add OpenAIProtocolLlmDriver in everruns-core as base protocol implementation
- Move AnthropicLlmDriver from everruns-core to everruns-anthropic crate
- Create OpenAILlmDriver in everruns-openai that wraps OpenAIProtocolLlmDriver
- Add register_driver functions to provider crates for startup registration
- Update CallModelAtom to use DriverRegistry instead of create_driver factory
- Add DriverNotRegistered error for graceful handling of missing drivers
- Rename llm_drivers.rs to llm_driver_registry.rs

Architecture:
- everruns-core: LlmDriver trait, DriverRegistry, OpenAIProtocolLlmDriver (base)
- everruns-openai: OpenAILlmDriver (production, wraps base for future extensions)
- everruns-anthropic: AnthropicLlmDriver (full implementation)
