+++ crates/everruns-core/Cargo.toml
@@ -60,6 +60,9 @@ openapi = ["utoipa"]
 tokio = { workspace = true, features = ["full", "test-util"] }
 tracing-subscriber.workspace = true
 wiremock = "0.6"
+# Provider crates for examples (dev-only, not a runtime dependency)
+everruns-openai = { path = "../everruns-openai" }
+everruns-anthropic = { path = "../everruns-anthropic" }
 
 [[example]]
 name = "decomposed_execution"
